refactor: melhorando mapperService
