+++ financial-control-backend/src/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { IncomeModule } from './income/income.module';
-import { IncomeRepository } from './income/income.Repository';
 import { PrismaModule } from './misc/prisma/prisma.module';
+import { IncomeRepository } from './income/repository/income.repository';
 
 @Module({
   imports: [IncomeModule,PrismaModule],
