+++ financial-control-backend/src/income/income.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { IncomeController } from './income.controller';
 import { IncomeService } from './income.service';
 import { MapperModule } from 'src/misc/mapper/mapper.module';
-import { IncomeRepository } from './income.Repository';
+import { IncomeRepository } from './repository/income.repository';
 
 @Module({
   imports: [MapperModule],
